Enable metadata writing for EventDisplay EventService.  ATEAM-832
Enable metadata writing for SP EventService (like in P1). ATEAM-832

SingleProcess Athena EventService mode used in EventDisplay was not
writing metedata to the fragment output files. This MR changes the
conditions tested in the OutputStream MetaDataStop incident to write
metadata in SP EventService ( MT EventService is not using incidents for
that)