Add possibility to update the RoiDescriptors to the view creation
Add possibility to update the RoiDescriptors to the view creation

Adds a new RoiUpdaterTool which can be included by the ViewCreator Tools.
This is implemented here in the Jet slice as a test. In principle, the
ToolHandle for the RoiUpdater Tool can be added to the ViewCreator base
class, and just not initialised.