DynVarFixerAlg Update, master branch (2022.06.01.)
Made it possible to use xAODMaker::DynVarFixerAlg with non-AuxContainerBase types.

Previously the algorithm would only work with auxiliary store types deriving
from xAOD::AuxContainerBase. But that's not the only auxiliary store type that
may need fixing in some circumstances.

While moving to using the auxiliary stores through their SG::IConstAuxStore
and SG::IAuxStoreIO interfaces, also migrated the algorthm to using
StoreGate handles. So that it could potentially be used in MT jobs as well.

Finally, added a simple jobO to the package that I used for testing the
algorithm.