GeoSpecialShapes: cleanup debug printout in LArWheelCalculator
The `fill_sincos_parameterization` of `LArWheelCalculator_Impl` had some
(partially) disabled code for debug printing. But some of the code was
still compiled and executed even during regular builds. Put the entire
section behind the `DEBUGPRINT` macro. This is also what the FIXME
comment indicated to be the correct thing to do.