TriggerMenuMT: replace/remove uses of configurableRun3Behavior
TriggerMenuMT: replace/remove uses of configurableRun3Behavior

To decide if we are running legacy or CA-based job configuration, it is
better to use `ComponentFactory.isRun3Cfg()` instead of
`configurableRun3Behavior`. This avoids the need to explicitly having to
set `configurableRun3Behavior` and also ensures coherency with what type
of configurable the `CompFactory` creates.