Trig jet monitoring - replace four concrete AlgTools used for monitoring L1...
Trig jet monitoring - replace foyr concrete AlgTools used for mnonitoring L! jets by a sigle templated AlgTool

On previous attemps to template the AlgTool on the jet container resulted in container version informaion
appearing in class names, and in the python configuration code, which is not acceptable.

This code encapsulates the xAOD constainer type in a struct container only a typedef. Using
such structs conceals the version information form the configuration code.