AthenaConfiguration: merge AthAlgSeq into TopAlg for legacy support
When using CA generators in legacy job options (e.g. via
`CAtoGlobalWrapper`) the list of algorithms needs to be appended to the
`TopAlg` sequence to ensure correct order with the old-style fragments
that populate that sequence directly.

Note that the `AppMgr` merges `TopAlg` into `AthAlgSeq` as the final
step in the configuration.