TrigValSteering: workaround to ignore false errors in AODtoDAOD logs
TrigValSteering: workaround to ignore false errors in AODtoDAOD logs

Hello,

This MR implements a workaround to ignore false positives in trigger tests during log file parsing.
The warning is known to be hamless in AODtoDAOD and should be ignored (comes from outdated CP smart slimming lists that don't reflect the R22 AOD content).
The typical warning is "xAOD::AuxSelection WARNING Selected dynamic Aux atribute XX not found in the registry".
We just alter one letter (dynamic -> Dynamic) so that the parsing ignores the warning later on.
It is done when the log files are all merged into one.

Relevant for ATR-25553, ATR-25449, maybe more.

Cheers,
Bertrand