Fixes for Truth derivations in master
Fixes for Truth derivations in master

Three fixes to get truth derivations running in master.

First, change the import of the EventDensity algorithm, which had been
fixed for the new config but not for the old config.

Second, turn off geometry loading when running truth derivations (it
isn't needed, and brings in a bunch of tools that also aren't needed).

Third, make sure that when a MET configuration is requested, if it's a
truth configuration it doesn't set up a bunch of track extrapolation
tools (which as far as I know, are not needed for such configurations).
If the suffix is something other than 'Truth' here (which should only be
the vanilla truth MET), then the tools are still loaded, which should
allow people to construct whatever slightly weird things they might
want.