Setting trigger matching prefix to TrigMatch_ in CA version of DAOD building
Setting trigger matching prefix to TrigMatch_ in CA version of DAOD building

In the course of the migration of the derivation framework to the CA, the prefix for the trigger matching containers was changed to `PhysCommonTau` (or `PhysCommonNoTau`) from `TrigMatch_`. Whilst not technically a bug this will be confusing for analysts and implies that there can be overlap between the tau and no-tau trigger lists, which there is not. Also it messes up the categorisation of the different container types in `checkxAOD.py` because the container name doesn't contain `Trig`. So with this MR, the prefix is reverted to `TrigMatch_`.

Since this only impacts the CA version of the derivation framework, which isn't yet in use, this has no impact on any ongoing activities.