Update muon trigger configuration of segment filter alg
Update muon trigger configuration of segment filter alg

Updating the configuration of the segment filtering algorithm to keep segments even if they are the same as the pre-filtered segments, and to drop the filtering based on the station (which was dropping all barrel and end-cap inner segments). With this the filtering will filter only based on the technology (STGC or MM), which is the desired behaviour for the trigger. This only affects the trigger when running in the mode to ignore the NSW information (ATR-25631).