TrigInDetConfig: disable TRT PID for tau signature
TrigInDetConfig: disable TRT PID for tau signature

Hello,

As discussed in ATR-25138, this MR is reverting the TRT PID settings to False for the tau trigger signature (were enabled in !51558).
This is because we currently don't exploit this new feature in tau triggers.
However we still foresee a possible use in tau ID, later this year, so the `eProbabilityNN` variable has been kept in TriggerEDMRun3.py, such that when we will enable TRT PID via a preExec in future AODFULL MC performance samples, the variable will naturally be in the AOD output.

Cheers,
Bertrand