Fix missing contents in CA version of DAOD_PHYS
Fix missing contents in CA version of DAOD_PHYS

This MR fixes two issues that led to variables being dropped from the CA version of DAOD_PHYS (and the other common physics content):

(1) Standard treatment of EventInfo in `OutputStreamConfig.py` caused a number of AuxDyn branches being set by the SlimmingHelper and then passed to OutputStreamConfig to be dropped. A new line has been added to skip the standard treatment of EventInfo and use the contents lists unaltered from the SlimmingHelper, if the job is a derivation production;
(2) Typo (missing `xAOD::` in a condition) led to truth particle and truth jet links being omitted from `TauJets`

Since this only impacts the CA version of the derivation framework which is not yet in use, this MR has no impact on any ongoing production.