AtlasFieldMapCondAlg: apply LoadMapOnStart only when running HLT
Move the `AtlasFieldMapCondAlg.LoadMapOnStart` setting into HLT-specific
job options. This should only be enabled when running in athenaHLT or in
a partition. Plain athena will not have a valid `EventContext` available
before the first event.

This went unnoticed until gaudi/Gaudi!1347 fixed a bug in the initialization
of the `EventIDRange` class.