Correctly configure track thinning in CA based DAOD building
Correctly configure track thinning in CA based DAOD building

Due to a mistake during the migration of the derivation framework to the component accumulator, a track thinning tool was not passed an instance of the track selector tool and so the thinning didn't take effect. This MR fixes that issue. Also an unrelated obsolete comment line is removed.