TrigTauHypo: minor cleanup of hypo tool
TrigTauHypo: minor cleanup of hypo tool

Hello,

This MR is doing minor cleanup, removing properties that are not used or always set to the same (default) value for all tau chains, therefore not really needed.
I also removed a file which apparently is used nowhere.
No change in counts is expected.

Cheers,
Bertrand