Remove truth from track overlay
Remove truth from track overlay

At the moment the pileup truth information is not stored in the RDO_BKG file so there is no point in trying to build the truth before the reconstruction stage.
This mainly removes the configuration options that allowed the truth containers to be copied through.
Additionally, the modifications to the InDetDetailedTrackTruthMaker that let it skip making the truth for some tracks are removed.