Cleanup/updating derivation ARTs for master and AthDerivation,22.2
Cleanup/updating derivation ARTs for master and AthDerivation,22.2

This MR removes all ARTs that are not needed in r22 (since the corresponding formats can use PHYS) and updates the non-PHYS that are already migrated to r22. The tests for TRUTH formats have been migrated to use already the Component Accumulator. The ARTs currently target also the new branch (built from master) AthDerivation,22.2 that we have recently created. I think we can run the tests only in master or 22.2,AthDerivation (since they are identical), but for this first MR I enabled both as check.