Fix TrigMuonEFHypoAlg to prevent ignoring inside-out muons
Fix TrigMuonEFHypoAlg to prevent ignoring inside-out muons

Adding some protection to TrigMuonEFHypoAlg to skip SA muons when we want to look only at combined muons, as we were sometimes only passing the SA muons to the hypo tool (where they were correctly rejected) and skipping over inside-out muons.

Also changing the check on the matching to previous decision for inside-out muons to always use the dR matching, as checking for identical extrapolated tracks doesn't work in this case.

(Also adding the muon author to the debug information about the muons, as it's usually useful when debugging).