Update muon idperf chains to not require combined muons
Update muon idperf chains to not require combined muons

The fixes in https://gitlab.cern.ch/atlas/athena/-/merge_requests/54498/ uncovered that the idperf chains were relying on the bug that was fixed. Adding a new menu sequence for the idperf chains that allow the hypo alg for combined muons to include SA muons so that combined muons are not implicitly required (which could bias the id performance results).