MMTriggerTool: reduce tool verbosity
MMTriggerTool: reduce tool verbosity

Hello,

This MR is downgrading a couple of WARNING to DEBUG, because these are printed out systematically a couple of times for every event, which is not useful.
In case the warnings indicate a real problem, then it has to be fixed by experts :-)
Bonus: this will make RDOtoRDOTrigger log files smaller.

Adding the urgent flag because it would be very desirable to have this in the MC21a release to be built imminently.

Cheers,
Bertrand