Declare conditions data container as mixed.
The BunchCrossingCondData depends on conditions data which uses both run/LB
and time stamp based IOVs. Thus the container has to be declared as mixed.

Also ensure that the bunch crossing conditions data is added with valid
timestamp and run/lumi ranges. Otherwise it cannot be added to the mixed
conditions container.