introduce new test files with DAOD_PHYS (ATLASG-1636)
This allows testing that software works on DAOD_PHYS.

I'm keeping the old files around with the suffix _OLD, as users have
specifically requested to be able to keep tests on older files which
may require different tool settings.

I also updated a fair number of the CP algorithm tests so that they
now explicitly use the _OLD files until issues are resolved.