Ensuring truth filter output is included in PHYS/LITE derivations
Ensuring truth filter output is included in PHYS/LITE derivations

These filters are important for being able to combine inclusive samples
(e.g. ttbar) with slices samples (MET or HT-sliced) correctly in
analysis. There are more elegant ways that are in development in PMG,
but for the time being this seems to be the best way and has been
validated by the SUSY group extensively.

It seems in the migration from 21.2 to master, these filters were
disabled on PHYS and PHYSLITE. This re-enables them.