CaloTPCnv: Fix roundtrip issue with CaloClusterCellLinkContainer.
CaloTPCnv: Fix roundtrip issue with CaloClusterCellLinkContainer.

When we save instances of CaloClusterCellLinkContainer, weights that are
exactly one are suppressed.  However, the container holds the weights
as double-precision, while what we save is single-precision.
So it's possible if a weight is very close to 1.0 but not exactly
equal to it, it gets saved as exactly 1.0 in single-precision.
So if we read and rewrite, the persistent form will change.
Avoid this by converting the weight to single-precision before
comparing to 1.0.
See ATLASRECTS-7129.