TrigAnalysisTest: update ref_RDOtoRDOTrig_v1DevHI_build
TrigAnalysisTest: update ref_RDOtoRDOTrig_v1DevHI_build

Hello,

This MR is updating a reference file used by a trigger ART test.
The test is timing out on the grid (ATR-25603), but locally it completes fine and complains about a count mismatch (which should be fixed by this MR).
The timeout error is not addressed.

Cheers,
Bertrand