AthDerivation Updates, master branch (2022.07.08.)
Taught the module how to handle not all packages being available.

This module is imported very early in AthDerivation jobs. In order
to avoid having to add all referenced packages to that project, the
module now expects the non-core packages to possibly not be
available. Giving INFO messages about every package/module not found.