tauMonitoring: fix lead track impact parameter and TRT PID variables
tauMonitoring: fix lead track impact parameter and TRT PID variables

Hello,

This MR is fixing a couple of variables in the offline tau DQ monitoring code, after looking at some Run3 data webdisplays.
The first is the retrieval of the d0 and z0 impact parameter significance of the leading track.
In R22, we don't have dedicated "leadTrack" variables anymore, because these are computed for all tracks.
The second is a small bug in the manipulation of TRT PID variables, currently they are overriden and set to 0.

Adding the 'urgent' flag in case it case be picked up for 22.0.76 expected tomorrow.

Cheers,
Bertrand