AthenaMonitoringKernel: deregistration fixes for LB histograms
Implement proper thread-safe cleanup of old LB histograms. The previous
code only kept one histogram alive, which is too short with a
publication interval that is larger than the lumiblock (ADHI-4905).

Now we are keeping histograms alive that belong to a fixed number of
lumiblocks (currently set to 5).

Relates to ADHI-4905, ATLASDQ-780, ATR-25564.