AthenaServices: Work around conditions dependency issues.
AthenaServices: Work around conditions dependency issues.

Some conditions objects have pointers to parts of other
conditions objects, which violates the lifetime guarantees of
conditions containers (a pointer you get from a conditions container
is guaranteed to be valid until the end of the current event, but
not past that).  In some cases, this can be dealt with by replacing
the pointers with CondLink, but that is sometimes rather inconvenient.
Try to work around this for now by ensuring that when we delete an object,
we also try to clean the containers of other objects that may depend
on it.