Monitor the b-jet preselection tracking in the TrigInDetAnalysis
Monitor the b-jet preselection tracking in the TrigInDetAnalysis

Currently have no tests which run this tracking, so this adds some chains which run the
preselection tracking, and adds the relevant analysis and plot making stages.

Currently this doesn;t producxe any output due to an issue for running the correct chains
in the menu, but these changes are useful and should be added so that we can concentrate
on getting the job working.