InDetCaloClusterROISelector: Fix potential vector overrun.
InDetCaloClusterROISelector: Fix potential vector overrun.

ROIPhiRZContainer::addROI can actually make two entries in the vector,
if the ROI is close to +-pi.  But if this happens. the length we've
set for max_output in CaloClusterROIPhiRZContainerMaker::execute
may be too small, leading to writes past the end of the vector.
We print a warning in that case, but by then it's too late.

Resize max_output if needed, and remove the warning, since it can
normally happen.  (Leave it commented out for documentation.)

See ATLASRECTS-7160.