Minor changes, protect against FPE in the ID Trigger monitoring
Minor changes, protect against FPE in the monitoring

In ATR-25821 it is reported of crashes in the IDTrigger monitoring when running
in multithreaded mode - although that may not be the fault to the ID Trigger monitoring.

As prt of this it was noticed the there were some dive by zero FPE due to corrupted
va;ues of qOverP which were 0, so fgiving rise to the FPE.

This protects against the division, but does not address what upstrwam code is producing
or corrupting the data.

It seems to run in the more recently nightlies, although the reasons for the original
crash are not yet understood.